fix(ssr): guard theme style injection against non-browser environments
